Revert 191584 "Build autofill_browser and autofill_renderer as s..."
Broke 64 bit windows build:
http://build.chromium.org/p/chromium.win/builders/Win%20x64%20Builder/builds/2883/steps/compile/logs/stdio
c:\b\build\slave\cr-win-rel-x64\build\src\components\autofill\renderer\password_autofill_manager.cc(580) :error C2220: warning treated as error - no 'object' file generated


> Build autofill_browser and autofill_renderer as standalone libs.
> 
> BUG=140037
> 
> TBR=ben@chromium.org, joi@chromium.org
> 
> Review URL: https://chromiumcodereview.appspot.com/12984015

TBR=kaiwang@chromium.org
Review URL: https://codereview.chromium.org/13261011

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@191587 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 9b2a64a081364b1e97c2ab0211a72986c3872650

+++ components/autofill/renderer/form_cache.cc
@@ -197,7 +197,7 @@ bool FormCache::ClearFormWithElement(const WebInputElement& element) {
       // Clearing the value in the focused node (above) can cause selection
       // to be lost. We force selection range to restore the text cursor.
       if (element == *input_element) {
-        int length = static_cast<int>(input_element->value().length());
+        int length = input_element->value().length();
         input_element->setSelectionRange(length, length);
       }
     } else if (IsSelectElement(control_element)) {
